mqtt_and_kafka/mqtt-router/requirements.txt
paho-mqtt==2.1.0
confluent-kafka>=2.4
